Needed refactor after merge.
