+++ src/main/java/hu/oe/nik/szfmv17t/automatedcar/AutomatedCar.java
@@ -3,7 +3,7 @@ package hu.oe.nik.szfmv17t.automatedcar;
 
 import hu.oe.nik.szfmv17t.automatedcar.bus.VirtualFunctionBus;
 import hu.oe.nik.szfmv17t.automatedcar.powertrainsystem.PowertrainSystem;
-import hu.oe.nik.szfmv17t.environment.WorldObject;
+import hu.oe.nik.szfmv17t.environment.domain.WorldObject;
 
 public class AutomatedCar extends WorldObject{
