--- src/main/java/hu/oe/nik/szfmv17t/environment/domain/Collidable.java
@@ -1,38 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package hu.oe.nik.szfmv17t.environment.domain;
-
-/**
- *
- * @author Gellert Babel <OE-NIK>
- */
-public class Collidable extends WorldObjectBase {
-    private float speed;
-
-    public void setSpeed(float speed) {
-        this.speed = speed;
-    }
-
-    public float getSpeed() {
-        return speed;
-    }
-
-    @Override
-    public void updateWorldObject() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public int getZIndex() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public WorldObjectState state() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-    
-}
