--- src/main/java/hu/oe/nik/szfmv17t/environment/domain/NonCollidable.java
@@ -1,30 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package hu.oe.nik.szfmv17t.environment.domain;
-
-/**
- *
- * @author Gellert Babel <OE-NIK>
- */
-//Base of stationary objects
-public class NonCollidable  extends WorldObjectBase{
-
-    @Override
-    public void updateWorldObject() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public int getZIndex() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public WorldObjectState state() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-    
-}
