--- src/main/java/hu/oe/nik/szfmv17t/environment/domain/Walker.java
@@ -1,14 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package hu.oe.nik.szfmv17t.environment.domain;
-
-/**
- *
- * @author Gellert Babel <OE-NIK>
- */
-public class Walker extends Collidable{
-    
-}
